+++ toolchains/rust/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+#### 🚀 Updates
+
+- Updated with moon v2 plugin APIs.
+
 ## 1.0.1
 
 #### 🐞 Fixes