+++ tools/bun/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Unreleased
+
+#### 🚀 Updates
+
+- Updated to support proto v0.55 release.
+- Added `pin_version` and `unpin_version` support, which maps to `package.json` `devEngines`.
+
+#### 🐞 Fixes
+
+- Fixed some `package.json` version parsing issues.
+
 ## 0.16.5
 
 #### 🐞 Fixes